backend/requirements.txt
fastapi>=0.109.0
uvicorn==0.24.0
motor==3.3.1
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==1.10.15
email-validator==2.1.0
